[REF] stock_multi_warehouse_security:
This is not a simple migration bettween 12.0 and 14.0
behaviour changes that users are able to only see warehouses set
on the user form or all if not set.

This module won't add more groups but based on ir.rules restrict
allowed data to see

Also we split ir.rules and the fact to add warehouse_id on models
in different modules.